Fix hard to track down bug
We were occasionally seeing bits of input disappearing from the final result.
This turned out to be due to regular parts of the input text being treated as
escape sequences.

We were checking for the presence of an escape character before calling
next-csi, but only once, instead of for each iteration, potentially causing a
second or third chunk of the same input to be treated as having an escape
sequence, even though it didn't.